Move saveFixSettings to common directory so it can be shared
Also update it to accept the fix container directly instead of an event object
